test: some attestation tests

+++ ucan/ucan.go
@@ -44,7 +44,7 @@ type Link = ipld.Link
 type Version = string
 
 // UTCUnixTimestamp is a timestamp in milliseconds since the Unix epoch.
-type UTCUnixTimestamp = uint64
+type UTCUnixTimestamp = int
 
 // https://github.com/ucan-wg/spec/#324-nonce
 type Nonce = string
